+++ CHANGELOG.md
@@ -426,3 +426,4 @@ Version 0.2.30 (TBD)
 --------------------
 
 * Add explicit mappings for <=, >=, >, !=
+* Make compound scalar objects match UnNamedCompoundScalar objects with the same properties.